+++ packages/manifold/contracts/lazyclaim/ILazyPayableClaim.sol
@@ -99,6 +99,6 @@ interface ILazyPayableClaim {
      * @param nonce                     the nonce that was signed
      * @param mintFor                   the address to mint for
      */
-    function mintSignature(address creatorContractAddress, uint256 instanceId, uint16 mintCount, bytes calldata signature, bytes32 message, bytes32 nonce, address mintFor, uint expiration) external payable;
+    function mintSignature(address creatorContractAddress, uint256 instanceId, uint16 mintCount, bytes calldata signature, bytes32 message, bytes32 nonce, address mintFor, uint256 expiration) external payable;
 
 }
